Added more tests

+++ Program/CCode/Testing/src/verifyOutputTest.c
@@ -15,10 +15,10 @@ TEST_TEAR_DOWN(VerifyOutput){
 TEST(VerifyOutput, testVerifyOutput1){
     struct parameters params;
     params = load_params("test.in");
-    float tempW[4] = {40, 42, 44, 46};
-    float tempP[4] = {40, 41.9, 43.8, 45.7};
-    float eW[4] = {0, 1000, 2000, 19800};
-    float eP[4] = {0, 1000, 2000, 5400};
+    double tempW[4] = {40, 42, 44, 46};
+    double tempP[4] = {40, 41.9, 43.8, 45.7};
+    double eW[4] = {0, 1000, 2000, 19800};
+    double eP[4] = {0, 1000, 2000, 5400};
     int sizeOfResults = sizeof(tempW) / sizeof(tempW[0]);
     int warnings = verify_output(tempW, tempP, eW, eP, params, sizeOfResults);
     TEST_ASSERT_EQUAL_INT(0, warnings);
@@ -27,10 +27,10 @@ TEST(VerifyOutput, testVerifyOutput1){
 TEST(VerifyOutput, testVerifyOutput2){
     struct parameters params;
     params = load_params("test.in");
-    float tempW[4] = {40, 42, 44, 46};
-    float tempP[4] = {40, 41.9, 43.8, 45.7};
-    float eW[4] = {0, 1000, 2000, 19800};
-    float eP[4] = {0, 1000, 2000, 3000};
+    double tempW[4] = {40, 42, 44, 46};
+    double tempP[4] = {40, 41.9, 43.8, 45.7};
+    double eW[4] = {0, 1000, 2000, 19800};
+    double eP[4] = {0, 1000, 2000, 3000};
     int sizeOfResults = sizeof(tempW) / sizeof(tempW[0]);
     int warnings = verify_output(tempW, tempP, eW, eP, params, sizeOfResults);
     TEST_ASSERT_EQUAL_INT(2, warnings);
@@ -39,10 +39,10 @@ TEST(VerifyOutput, testVerifyOutput2){
 TEST(VerifyOutput, testVerifyOutput3){
     struct parameters params;
     params = load_params("test.in");
-    float tempW[4] = {40, 42, 44, 46};
-    float tempP[4] = {40, 41.9, 43.8, 45.7};
-    float eW[4] = {0, 1000, 2000, 3000};
-    float eP[4] = {0, 1000, 2000, 5400};
+    double tempW[4] = {40, 42, 44, 46};
+    double tempP[4] = {40, 41.9, 43.8, 45.7};
+    double eW[4] = {0, 1000, 2000, 3000};
+    double eP[4] = {0, 1000, 2000, 5400};
     int sizeOfResults = sizeof(tempW) / sizeof(tempW[0]);
     int warnings = verify_output(tempW, tempP, eW, eP, params, sizeOfResults);
     TEST_ASSERT_EQUAL_INT(1, warnings);
@@ -51,10 +51,10 @@ TEST(VerifyOutput, testVerifyOutput3){
 TEST(VerifyOutput, testVerifyOutput4){
     struct parameters params;
     params = load_params("test.in");
-    float tempW[4] = {40, 42, 44, 46};
-    float tempP[4] = {40, 41.9, 43.8, 45.7};
-    float eW[4] = {0, 1000, 2000, 3000};
-    float eP[4] = {0, 1000, 2000, 3000};
+    double tempW[4] = {40, 42, 44, 46};
+    double tempP[4] = {40, 41.9, 43.8, 45.7};
+    double eW[4] = {0, 1000, 2000, 3000};
+    double eP[4] = {0, 1000, 2000, 3000};
     int sizeOfResults = sizeof(tempW) / sizeof(tempW[0]);
     int warnings = verify_output(tempW, tempP, eW, eP, params, sizeOfResults);
     TEST_ASSERT_EQUAL_INT(3, warnings);
